Use react Router 6.4, created Home and new NavLink to switch from the TodoList and the New TodoList page
