+++ package.json
@@ -13,7 +13,8 @@
     "bootstrap": "^5.3.2",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.1",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "6.4"
   },
   "devDependencies": {
     "@types/react": "^18.2.55",
