+++ src/New.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export function New() {
+  return (
+    <div>New Todo List</div>
+  )
+}
